+++ app/src/main/java/edu/uw/tcss450/chatapp/ui/chat/ChatFragment.java
@@ -1,4 +1,4 @@
-package edu.uw.tcss450.chatapp;
+package edu.uw.tcss450.chatapp.ui.chat;
 
 import android.os.Bundle;
 
@@ -8,13 +8,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import edu.uw.tcss450.chatapp.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class ContactsFragment extends Fragment {
+public class ChatFragment extends Fragment {
 
-    public ContactsFragment() {
+    public ChatFragment() {
         // Required empty public constructor
     }
 
@@ -23,6 +25,6 @@ public class ContactsFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_contacts, container, false);
+        return inflater.inflate(R.layout.fragment_chat, container, false);
     }
 }
