+++ app/src/main/java/edu/uw/tcss450/chatapp/ui/contacts/ContactsFragment.java
@@ -1,4 +1,4 @@
-package edu.uw.tcss450.chatapp;
+package edu.uw.tcss450.chatapp.ui.contacts;
 
 import android.os.Bundle;
 
@@ -8,13 +8,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import edu.uw.tcss450.chatapp.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class SuccessFragment extends Fragment {
+public class ContactsFragment extends Fragment {
 
-    public SuccessFragment() {
+    public ContactsFragment() {
         // Required empty public constructor
     }
 
@@ -23,6 +25,6 @@ public class SuccessFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_success, container, false);
+        return inflater.inflate(R.layout.fragment_contacts, container, false);
     }
 }
