+++ app/src/main/java/edu/uw/tcss450/chatapp/ui/home/signin/LoginFragment.java
@@ -1,4 +1,4 @@
-package edu.uw.tcss450.chatapp;
+package edu.uw.tcss450.chatapp.ui.home.signin;
 
 import android.os.Bundle;
