+++ app/src/main/res/layout/fragment_chat.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout6"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ChatFragment" >
+    tools:context=".ui.chat.ChatFragment" >
 
     <TextView
         android:id="@+id/textView_chat_greeting"
