+++ app/src/main/res/layout/fragment_contacts.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout5"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ContactsFragment" >
+    tools:context=".ui.contacts.ContactsFragment" >
 
     <TextView
         android:id="@+id/textView_contacts_greeting"
