+++ app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginFragment">
+    tools:context=".ui.home.signin.LoginFragment">
 
     <Button
         android:id="@+id/button_login_register"
