+++ app/src/main/res/layout/fragment_register.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegisterFragment" >
+    tools:context=".ui.home.register.RegisterFragment" >
 
     <Button
         android:id="@+id/button_register_register"
