+++ app/src/main/res/layout/fragment_success.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout3"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SuccessFragment">
+    tools:context=".ui.home.SuccessFragment">
 
     <TextView
         android:id="@+id/textView_success_greeting"
