+++ app/src/main/res/layout/fragment_weather.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout4"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WeatherFragment" >
+    tools:context=".ui.weather.WeatherFragment" >
 
     <TextView
         android:id="@+id/textView_weather_greeting"
